Attempt at repairing issue 13 by no longer handing a RepairManager to the caller. Instead, the service allows you to ask it to run the repair manager given the supplied decision provider. The service may only run one repair manager at a time, calls for repair while the service is servicing a repair are not queued, they are ignored. This is because once the repair manager performs some events, the other requests are subsequently invalid, though to be honest, most of the time they'll contain the same decision provider of the call that is currently running.
The entries aren't logged because in the case of OSX, the can be quite a lot of them. Read the stack trace in issue 13 for a clarification of how in OSX, clicks get passed from the task repair menu dialog back to the trim widget, resulting in another copy of the dialog.